Print the size of data types

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 
 	printf("Size of a char: %u byte(s)\n", sizeof(c));
 	printf("Size of an int: %u byte(s)\n", sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned int)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long int)sizeof(i));
 	printf("Size of a long long int: %llu byte(s)\n", (unsigned long long int)sizeof(i));
 	printf("Size of a float: %u byte(s)\n", sizeof(f);
 	return (0);
